Implementation of veterinary records

+++ VetCare/src/VetCare/Control/VetJpaController.java
@@ -4,17 +4,20 @@
  */
 package VetCare.Control;
 
+import VetCare.Control.exceptions.IllegalOrphanException;
 import VetCare.Control.exceptions.NonexistentEntityException;
-import VetCare.Control.exceptions.PreexistingEntityException;
-import VetCare.Entities.Classes.Vet;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import VetCare.Modelo.Appointment;
+import VetCare.Modelo.Vet;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -31,18 +34,31 @@ public class VetJpaController implements Serializable {
         return emf.createEntityManager();
     }
 
-    public void create(Vet vet) throws PreexistingEntityException, Exception {
+    public void create(Vet vet) {
+        if (vet.getAppointmentCollection() == null) {
+            vet.setAppointmentCollection(new ArrayList<Appointment>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            em.persist(vet);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findVet(vet.getVetId()) != null) {
-                throw new PreexistingEntityException("Vet " + vet + " already exists.", ex);
+            Collection<Appointment> attachedAppointmentCollection = new ArrayList<Appointment>();
+            for (Appointment appointmentCollectionAppointmentToAttach : vet.getAppointmentCollection()) {
+                appointmentCollectionAppointmentToAttach = em.getReference(appointmentCollectionAppointmentToAttach.getClass(), appointmentCollectionAppointmentToAttach.getAppointmentId());
+                attachedAppointmentCollection.add(appointmentCollectionAppointmentToAttach);
             }
-            throw ex;
+            vet.setAppointmentCollection(attachedAppointmentCollection);
+            em.persist(vet);
+            for (Appointment appointmentCollectionAppointment : vet.getAppointmentCollection()) {
+                Vet oldVetIdOfAppointmentCollectionAppointment = appointmentCollectionAppointment.getVetId();
+                appointmentCollectionAppointment.setVetId(vet);
+                appointmentCollectionAppointment = em.merge(appointmentCollectionAppointment);
+                if (oldVetIdOfAppointmentCollectionAppointment != null) {
+                    oldVetIdOfAppointmentCollectionAppointment.getAppointmentCollection().remove(appointmentCollectionAppointment);
+                    oldVetIdOfAppointmentCollectionAppointment = em.merge(oldVetIdOfAppointmentCollectionAppointment);
+                }
+            }
+            em.getTransaction().commit();
         } finally {
             if (em != null) {
                 em.close();
@@ -50,17 +66,50 @@ public class VetJpaController implements Serializable {
         }
     }
 
-    public void edit(Vet vet) throws NonexistentEntityException, Exception {
+    public void edit(Vet vet) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Vet persistentVet = em.find(Vet.class, vet.getVetId());
+            Collection<Appointment> appointmentCollectionOld = persistentVet.getAppointmentCollection();
+            Collection<Appointment> appointmentCollectionNew = vet.getAppointmentCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Appointment appointmentCollectionOldAppointment : appointmentCollectionOld) {
+                if (!appointmentCollectionNew.contains(appointmentCollectionOldAppointment)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Appointment " + appointmentCollectionOldAppointment + " since its vetId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Appointment> attachedAppointmentCollectionNew = new ArrayList<Appointment>();
+            for (Appointment appointmentCollectionNewAppointmentToAttach : appointmentCollectionNew) {
+                appointmentCollectionNewAppointmentToAttach = em.getReference(appointmentCollectionNewAppointmentToAttach.getClass(), appointmentCollectionNewAppointmentToAttach.getAppointmentId());
+                attachedAppointmentCollectionNew.add(appointmentCollectionNewAppointmentToAttach);
+            }
+            appointmentCollectionNew = attachedAppointmentCollectionNew;
+            vet.setAppointmentCollection(appointmentCollectionNew);
             vet = em.merge(vet);
+            for (Appointment appointmentCollectionNewAppointment : appointmentCollectionNew) {
+                if (!appointmentCollectionOld.contains(appointmentCollectionNewAppointment)) {
+                    Vet oldVetIdOfAppointmentCollectionNewAppointment = appointmentCollectionNewAppointment.getVetId();
+                    appointmentCollectionNewAppointment.setVetId(vet);
+                    appointmentCollectionNewAppointment = em.merge(appointmentCollectionNewAppointment);
+                    if (oldVetIdOfAppointmentCollectionNewAppointment != null && !oldVetIdOfAppointmentCollectionNewAppointment.equals(vet)) {
+                        oldVetIdOfAppointmentCollectionNewAppointment.getAppointmentCollection().remove(appointmentCollectionNewAppointment);
+                        oldVetIdOfAppointmentCollectionNewAppointment = em.merge(oldVetIdOfAppointmentCollectionNewAppointment);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                String id = vet.getVetId();
+                Integer id = vet.getVetId();
                 if (findVet(id) == null) {
                     throw new NonexistentEntityException("The vet with id " + id + " no longer exists.");
                 }
@@ -73,7 +122,7 @@ public class VetJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -84,6 +133,17 @@ public class VetJpaController implements Serializable {
                 vet.getVetId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The vet with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<Appointment> appointmentCollectionOrphanCheck = vet.getAppointmentCollection();
+            for (Appointment appointmentCollectionOrphanCheckAppointment : appointmentCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Vet (" + vet + ") cannot be destroyed since the Appointment " + appointmentCollectionOrphanCheckAppointment + " in its appointmentCollection field has a non-nullable vetId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(vet);
             em.getTransaction().commit();
@@ -118,7 +178,7 @@ public class VetJpaController implements Serializable {
         }
     }
 
-    public Vet findVet(String id) {
+    public Vet findVet(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Vet.class, id);
